django-quasar-railway-blueprint/backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
dj-database-url==2.1.0
psycopg2-binary==2.9.9
python-decouple==3.8
whitenoise==6.6.0
drf-spectacular==0.27.2
gunicorn==21.2.0